requirements.txt
requests
fastapi
uvicorn
python-multipart
python-dotenv
google-generativeai
langchain
langchain-community
chromadb
sentence-transformers
PyPDF2
numpy